split packages/accel-record-core/src/search/index.ts
